- Made read/write functions use ssize_t
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@8957 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/send.c
@@ -171,7 +171,7 @@ sendq_unblocked(fde_t *F, void *data)
 void
 send_queued_write(struct Client *to)
 {
-  int retlen = 0;
+  ssize_t retlen = 0;
 
   /*
    ** Once socket is marked dead, we cannot start writing to it,
